Changed the way the app fetches movie counter

+++ app/src/main/java/fiveagency/internship/food/movieapp/ui/profile/ProfileFragment.kt
@@ -38,7 +38,7 @@ class ProfileFragment : BaseFragment<ProfileContract.Presenter>(), ProfileContra
         presenter.start()
     }
 
-    override fun render(displayName: String, email: String, movieCount: Int, favoritesCount: Int) {
+    override fun render(displayName: String, email: String, movieCount: Long, favoritesCount: Int) {
         userDisplayNameTextView?.text = displayName
         userEmailTextView?.text = email
         movieCountTextView?.text = movieCount.toString()
